# perf: Streamline leverage CLI tests