@toteki 2022-05-30
I also tried renaming the two files to end with `_test.go` but the suite consistently errors in that case

```
    suite_test.go:36: 
        	Error Trace:	suite_test.go:36
        	            				suite.go:118
        	            				cli_test.go:45
        	Error:      	Received unexpected error:
        	            	timeout exceeded waiting for block
        	Test:       	TestIntegrationTestSuite
```
So for now, just leaving the coverage situation as it was.